fetching from weather api vf with time range limit

+++ apiweathercore/src/main/java/com/apiweather/app/biz/model/weather/AgroForcast.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 @Setter
 public class AgroForcast {
 	private Long stationId;
-	private int dt;
+	private long dt;
     private List<WeatherCondition> weather;
     private Main main;
     private Wind wind;
