+++ apiweathercore/src/main/java/com/apiweather/app/rest/clients/WeatherApiCaller.java
@@ -8,7 +8,7 @@ import com.apiweather.app.biz.model.weather.AgroForcast;
 public interface WeatherApiCaller {
 
 	
-	public List<AgroForcast> getForecastByLatLong (long lon, long lat);
+	public List<AgroForcast> getForecastByLatLong (double lon, double lat);
 	
-	public AgroWeather getCurrentWeatherByLatAndLong(long lon, long lat);
+	public AgroWeather getCurrentWeatherByLatAndLong(double lon, double lat);
 }
